requirements.txt
quart
aiohttp
quart-cors
python-dotenv
beautifulsoup4